Execute events on run loop in Linux
